Calcul fait en 32 bits et non en 64 bits
(cherry picked from commit 02ea75af52b584a3e47bd6b96a12762b3ea3eecf)

+++ src/drive/tape_drive/util/scsi.c
@@ -39,6 +39,8 @@
 #include <scsi/sg.h>
 // bool
 #include <stdbool.h>
+// uint*_t
+#include <stdint.h>
 // snprintf
 #include <stdio.h>
 // free
@@ -1401,7 +1403,7 @@ int sodr_tape_drive_scsi_size_available(int fd, struct so_media * media) {
 	if (status || result_header->page_code != 0x17)
 		return 1;
 
-	unsigned long long free_block = 0, total_block = 0;
+	uint64_t free_block = 0, total_block = 0;
 
 	result_header->page_length = be16toh(result_header->page_length);
 	unsigned int position;
@@ -1414,42 +1416,42 @@ int sodr_tape_drive_scsi_size_available(int fd, struct so_media * media) {
 			case 0x0202:
 				vsl->value.partition_log[0].partition_number = be16toh(vsl->value.partition_log[0].partition_number);
 				vsl->value.partition_log[0].int32 = be32toh(vsl->value.partition_log[0].int32);
-				total_block = vsl->value.partition_log[0].int32 << 20;
+				total_block = ((uint64_t) (vsl->value.partition_log[0].int32)) << 20;
 
 				if (vsl->parameter_length < 16)
 					break;
 
 				vsl->value.partition_log[1].partition_number = be16toh(vsl->value.partition_log[1].partition_number);
 				vsl->value.partition_log[1].int32 = be32toh(vsl->value.partition_log[1].int32);
-				total_block += vsl->value.partition_log[1].int32 << 20;
+				total_block += ((uint64_t) (vsl->value.partition_log[1].int32)) << 20;
 
 				break;
 
 			case 0x0203:
 				vsl->value.partition_log[0].partition_number = be16toh(vsl->value.partition_log[0].partition_number);
 				vsl->value.partition_log[0].int32 = be32toh(vsl->value.partition_log[0].int32);
-				free_block = total_block - (vsl->value.partition_log[0].int32 << 20);
+				free_block = total_block - (((uint64_t) (vsl->value.partition_log[0].int32)) << 20);
 
 				if (vsl->parameter_length < 16)
 					break;
 
 				vsl->value.partition_log[1].partition_number = be16toh(vsl->value.partition_log[1].partition_number);
 				vsl->value.partition_log[1].int32 = be32toh(vsl->value.partition_log[1].int32);
-				free_block = total_block - (vsl->value.partition_log[1].int32 << 20);
+				free_block = total_block - (((uint64_t) (vsl->value.partition_log[1].int32)) << 20);
 
 				break;
 
 			case 0x0204:
 				vsl->value.partition_log[0].partition_number = be16toh(vsl->value.partition_log[0].partition_number);
 				vsl->value.partition_log[0].int32 = be32toh(vsl->value.partition_log[0].int32);
-				free_block = vsl->value.partition_log[0].int32 << 20;
+				free_block = ((uint64_t) (vsl->value.partition_log[0].int32)) << 20;
 
 				if (vsl->parameter_length < 16)
 					break;
 
 				vsl->value.partition_log[1].partition_number = be16toh(vsl->value.partition_log[1].partition_number);
 				vsl->value.partition_log[1].int32 = be32toh(vsl->value.partition_log[1].int32);
-				free_block = vsl->value.partition_log[1].int32 << 20;
+				free_block = ((uint64_t) (vsl->value.partition_log[1].int32)) << 20;
 
 				break;
 		}
